+++ relbench/base/__init__.py
@@ -2,8 +2,8 @@ from .database import Database
 from .dataset import Dataset
 from .table import Table
 from .task_base import BaseTask, TaskType
-from .task_link import RecommendationTask
-from .task_node import EntityTask
+from .task_entity import EntityTask
+from .task_recommendation import RecommendationTask
 
 __all__ = [
     "Database",
